move httperror to its own package
